Attach coordinate field to integral rather than measure
With UFL 1.3 the measure is no longer attached to the integral.
Instead the domain data is attached straight to the integral. Instead
of stashing the coordinate field as the domain data in the Measure,
we directly attach it the UFL element when building a FunctionSpace.

UFL form preprocessing will then make collate the domain data
according to its ufl_id, which is the uuid of the Function for the
coordinate field.